backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-docx==1.1.0
lxml==4.9.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic>=2.7.0
pydantic-settings>=2.0.0
python-multipart==0.0.6
aiofiles==23.2.1
pillow==10.1.0
python-dotenv==1.0.0
